refactor adding page to group into new hooks useAddPageToGroup
